Notification_Services/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
aiosmtplib
python-dotenv
orjson